move logic to dedicated validator module
